requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.0.0
streamlit==1.28.0
pandas==2.0.0
numpy==1.24.0
scikit-learn==1.3.0
joblib==1.3.0
openpyxl==3.1.0
python-decouple==3.8
requests==2.31.0
Pillow==10.0.0
plotly==5.17.0
